GNG-492: Move global Table component into Components directory as its not generic enough
